feat: V2 suspend state key to avoid partial wake in old versions
The feature to suspend cron jobs (will be released as v0.4.0)
is not backwards compatible with older versions.

Meaning a cronJob that was suspended with version v0.4.0 or above
can not be awoken with previous versions of kubesleep.

This commit makes this incompatibility explicit by explicitly failing
instead of leaving the cronJob in a suspended state.